+++ components/vf-blockquote/vf-blockquote.angular/lib/vf-blockquote.angular.module.d.ts
@@ -1,0 +1,8 @@
+import * as i0 from "@angular/core";
+import * as i1 from "./vf-blockquote.angular.component";
+import * as i2 from "@angular/common";
+export declare class VfBlockquoteAngularModule {
+    static ɵfac: i0.ɵɵFactoryDeclaration<VfBlockquoteAngularModule, never>;
+    static ɵmod: i0.ɵɵNgModuleDeclaration<VfBlockquoteAngularModule, [typeof i1.VfBlockquoteAngularComponent], [typeof i2.CommonModule], [typeof i1.VfBlockquoteAngularComponent]>;
+    static ɵinj: i0.ɵɵInjectorDeclaration<VfBlockquoteAngularModule>;
+}